Move FBX modules to subdirectory
